abstract out volume and audioElement logic
